auction create folder structure
