+++ src/views/createAuction/createAuction.scss
@@ -1,0 +1,42 @@
+@import "../../style/variables";
+
+.create-auction-wrapper {
+  height: calc(100vh - 80px);
+  background-color: #f2f0f1;
+  padding: 40px;
+
+  .auction-form-card {
+    width: 100%;
+    min-height: 650px;
+    padding: 40px;
+    border-radius: 16px;
+    background-color: #fff;
+  }
+
+  .auction-stepper-btn {
+    button {
+      background-color: none;
+      outline: none;
+      border: none;
+      width: 186px;
+      height: 52px;
+      padding: 16px;
+      border-radius: 62px;
+      color: #000;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      font-weight: 700;
+      font-size: 16px;
+      &:disabled {
+        cursor: not-allowed;
+        background-color: #f2f0f1;
+        opacity: 0.5;
+      }
+    }
+    .next-btn {
+      background-color: $primary-text;
+      color: #fff;
+    }
+  }
+}
